fix unittest after refactoring #159
